requirements.txt
pandas
numpy
psycopg2
python-dotenv
dbt-postgres
dbt-core
telethon